% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multiply_Number.R
\name{Multiply_Number}
\alias{Multiply_Number}
\title{Multiply by a number.}
\usage{
Multiply_Number(i, n)
}
\arguments{
\item{i}{Number.}

\item{n}{Number to multiply by}
}
\value{
List.  Output a named list.  Element \code{success} is a boolean denoting if the function was successful.
    Element \code{output} contains the function result (a numeric in the case of a success) or an error message
    (in the case of a failure)
}
\description{
Node to take the incoming number, multiply by a specified number and return the product
}
\examples{
\dontrun{
Multiply_Number(3, 5)
}

}

